+++ R/gt_make_raster.R
@@ -15,6 +15,8 @@
 #' @param webshot_zoom How many pixels should be created relative to height and width values. If `height` and `width` are set to `100` and `webshot_zoom` is set to `2`, the resulting raster will have dimensions of about `200x200` (default: `1`). 
 #' @param webshot_delay How long to wait for Google traffic layer to render. Larger height/widths require longer delay times. If `NULL`, the following delay time (in seconds) is used: `delay = max(height,width)/200`.
 #' @param print_progress Whether to print function progress (default: `TRUE`)
+#' @param cache_dir Where to save the temporary data. If `NULL`, using the system default temp dir (default: `NULL`)
+#' @param delete_cache Whether to delete the temporary data. (default: `TRUE`)
 #' 
 #' @references Markus Hilpert, Jenni A. Shearston, Jemaleddin Cole, Steven N. Chillrud, and Micaela E. Martinez. [Acquisition and analysis of crowd-sourced traffic data](https://arxiv.org/abs/2105.12235). CoRR, abs/2105.12235, 2021.
 #' @references Pavel Pokorny. [Determining traffic levels in cities using google maps](https://ieeexplore.ieee.org/abstract/document/8326831). In 2017 Fourth International Conference on Mathematics and Computers in Sciences and in Industry (MCSI), pages 144–147, 2017.
@@ -52,16 +54,28 @@ gt_make_raster <- function(location,
                            traffic_color_dist_metric = "CIEDE2000",
                            webshot_zoom = 1,
                            webshot_delay = NULL,
-                           print_progress = TRUE){
+                           print_progress = TRUE,
+                           cache_dir = NULL,
+                           delete_cache = TRUE){
   
   ## Set webshot_delay if null
   webshot_delay <- gt_estimate_webshot_delay(height, width, webshot_delay)
   
   ## Filename; as html
-  temp_dir <- tempdir() %>% str_replace_all("\\\\", "/")
-  filename_html <- tempfile(pattern = "file", tmpdir = temp_dir, fileext = ".html") %>%
-    str_replace_all("\\\\", "/")
-  
+  if (is.character(cache_dir)) {
+    dir.create(cache_dir, showWarnings = FALSE)
+    temp_dir <- cache_dir
+    filename_html <- file.path(temp_dir, "file.html")
+  } else {
+    temp_dir <- tempdir() %>% str_replace_all("\\\\", "/")
+    filename_html <- tempfile(pattern = "file", tmpdir = temp_dir, fileext = ".html") %>%
+      str_replace_all("\\\\", "/")
+  }
+
+  # temp_dir <- tempdir() %>% str_replace_all("\\\\", "/")
+  # filename_html <- tempfile(pattern = "file", tmpdir = temp_dir, fileext = ".html") %>%
+  #   str_replace_all("\\\\", "/")
+
   ## Make html
   gt_make_html(location = location,
                height = height,
@@ -83,7 +97,9 @@ gt_make_raster <- function(location,
                          print_progress = print_progress)
   
   ## Delete html file
-  unlink(filename_html)
-  
+  if (delete_cache) {
+    unlink(filename_html)
+  }
+
   return(r)
 }
